Created controller folders and moved teacher controllers into appropriate folder
