+++ src/main/resources/com/example/cab302tailproject/ai_assistant-teacher.fxml
@@ -12,7 +12,7 @@
             maxHeight = "-Infinity"
             prefHeight = "420"
             prefWidth="710"
-            fx:controller="com.example.cab302tailproject.controller.AiAssistantController" style= "-fx-background-color: white;">
+            fx:controller="com.example.cab302tailproject.controller.teachercontroller.AiAssistantController" style= "-fx-background-color: white;">
     <left>
         <VBox prefHeight="420" prefWidth="200" style="-fx-border-color: #000000; -fx-border-width: 2;">
             <VBox alignment="TOP_LEFT" prefHeight="120" style="-fx-padding: 25 0 0 30;">
